# Cython additions