--- a/.codecov.yml
+++ b/.codecov.yml
@@ -1,0 +1,4 @@
+flags:
+  project:
+    paths:
+      - sparse/