--- a/.codecov.yml
+++ b/.codecov.yml
@@ -1,4 +1,17 @@
+coverage:
+  precision: 2
+  round: down
+  range: 70...98
+
+  ignore:
+    - !sparse/
+
 flags:
   project:
-    paths:
-      - sparse/
+    enabled: yes
+    target: 95
+
+  patch:
+    default:
+      enabled: yes
+      target: 96