--- a/sparse/__init__.py
+++ b/sparse/__init__.py
@@ -2,4 +2,3 @@ from .coo import *
 from .dok import DOK
 from .sparse_array import SparseArray
 from .utils import random
-from ._version import __version__