--- a/sparse/coo/umath_cython.pyx
+++ b/sparse/coo/umath_cython.pyx
@@ -2,15 +2,15 @@ import numpy as np
 cimport numpy as np
 cimport cython
 from libcpp.vector cimport vector
-from numpy cimport int64_t
+from numpy cimport int64_t, uint64_t
 
 cdef extern from "<cstring>" namespace "std":
     void* memcpy (void* destination, const void* source, size_t num)
 
 
 def cywhere(np.ndarray a not None, np.ndarray b not None):
-    a = a.astype(dtype=np.int64)
-    b = b.astype(dtype=np.int64)
+    a = a.astype(dtype=np.uint64)
+    b = b.astype(dtype=np.uint64)
 
     a_ind, b_ind = cywhere_wrapped(a, b)
 
@@ -21,18 +21,19 @@ def cywhere(np.ndarray a not None, np.ndarray b not None):
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
-cdef cywhere_wrapped(int64_t[:] a, int64_t[:] b):
-    cdef vector[int64_t] a_ind = vector[int64_t]()
-    cdef vector[int64_t] b_ind = vector[int64_t]()
-    cdef int64_t na = a.shape[0], nb = b.shape[0]
-    cdef int64_t ia, ib
-    cdef int64_t j
-    cdef int64_t match = 0
+cdef cywhere_wrapped(uint64_t[:] a, uint64_t[:] b):
+    cdef vector[size_t] a_ind = vector[size_t]()
+    cdef vector[size_t] b_ind = vector[size_t]()
+    cdef size_t na = a.shape[0], nb = b.shape[0]
+    cdef size_t ia, ib = 0
+    cdef uint64_t j
+    cdef uint64_t match = 0
 
     with nogil:
         for ia in range(na):
             j = a[ia]
-            ib = match
+            if j == b[match]:
+                ib = match
             while ib < nb and j >= b[ib]:
                 if j == b[ib]:
                     a_ind.push_back(ia)
@@ -46,8 +47,8 @@ cdef cywhere_wrapped(int64_t[:] a, int64_t[:] b):
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
-cdef vector_to_np(vector[int64_t]& vec):
-    cdef np.ndarray[int64_t] arr = np.empty(vec.size(), dtype=np.int64)
+cdef vector_to_np(vector[size_t]& vec):
+    cdef np.ndarray[uint64_t] arr = np.empty(vec.size(), dtype=np.uint64)
 
     cdef void* arr_p = <void*> arr.data
     cdef void* vec_p = <void*> &vec[0]