… docs/conf.py
… version.py