~~Do NOT merge this PR yet.~~

Update: Now ready to merge.

I wrote up a simple algorithm to do what `_match_arrays` does in Cython (it takes up most of the time in `elemwise`). Unfortunately, I can't seem to be able to make it faster than the original. At best, it approaches the original. `--march=native` and `-O3` don't make a difference.

Take this as a conversation starter or a "suggestions?" PR to build upon.