@hameerabbasi 2018-03-03
This is now actually faster. I realized that I was resetting the index for `b` even when the next item _wasn't_ a match. This caused `ib` to jump backwards by large amounts sometimes, which was causing the slowdowns. To be fair, this was in the original code from SO.

Benchmarks:
Old:
```
In[7]: %timeit x + y
6.47 s ± 38.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
In[8]: %timeit x * y
1.91 s ± 21 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

New:
```
In[7]: %timeit x + y
4.14 s ± 37 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
In[8]: %timeit x * y
1.15 s ± 15.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```