@hameerabbasi 2018-03-03
With a short cursory look, it seems the majority of time is now being taken up by `argsort` (44/75 ms in a 5000x5000 `operator.add` with density 0.01)