@hameerabbasi 2018-03-03
I'm thinking of merging this since it's now faster by a significant margin. We would need to update installation docs to say you need a C++ compiler.

How would we manage distribution? A second `setup.py`?