@mrocklin 2018-03-03
Adding Cython is, I believe, a significant step for this project.  I will
probably want to discuss this further but may not have time to do so this
weekend.  If possible I'd like to wait before merging.

On Sat, Mar 3, 2018 at 1:52 PM, Hameer Abbasi <notifications@github.com>
wrote:

> I'm thinking of merging this since it's now faster by a significant
> margin. We would need to update installation docs to say you need a C++
> compiler.
>
> How would we manage distribution? A second setup.py?
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/pydata/sparse/pull/123#issuecomment-370170788>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AASszK18zNhVtquU3NgFzRsKKJt6qnSDks5tauZtgaJpZM4Sa4np>
> .
>