@hameerabbasi 2018-03-05
Closing until #126 is properly resolved. Whatever is decided on, I'll rewrite this in that language.